fix: organize multicall related functionality + tests in separate mod
