+++ crates/contract/src/lib.rs
@@ -30,6 +30,8 @@ pub use instance::*;
 mod call;
 pub use call::*;
 
+mod multicall;
+
 // Not public API.
 // NOTE: please avoid changing the API of this module due to its use in the `sol!` macro.
 #[doc(hidden)]
